# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API para Bot de Ventas WhatsApp con OpenClaw y Ollama
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: Listar contactos/clientes
      responses:
        "200":
          description: Lista de contactos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactList"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Crear nuevo contacto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactRequest"
      responses:
        "201":
          description: Contacto creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      summary: Obtener contacto por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Contacto
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    put:
      operationId: updateContact
      tags: [contacts]
      summary: Actualizar contacto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContactRequest"
      responses:
        "200":
          description: Contacto actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: Listar productos del catálogo
      responses:
        "200":
          description: Lista de productos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
    post:
      operationId: createProduct
      tags: [products]
      summary: Crear producto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Producto creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    put:
      operationId: updateProduct
      tags: [products]
      summary: Actualizar producto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        "200":
          description: Producto actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Eliminar producto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Producto eliminado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
  /automation/rules:
    get:
      operationId: listAutomationRules
      tags: [automation]
      summary: Listar reglas de automatización
      responses:
        "200":
          description: Reglas de automatización
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationRuleList"
    post:
      operationId: createAutomationRule
      tags: [automation]
      summary: Crear regla de automatización
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAutomationRuleRequest"
      responses:
        "201":
          description: Regla creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationRule"
  /automation/rules/{id}:
    put:
      operationId: updateAutomationRule
      tags: [automation]
      summary: Actualizar regla de automatización
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAutomationRuleRequest"
      responses:
        "200":
          description: Regla actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationRule"
    delete:
      operationId: deleteAutomationRule
      tags: [automation]
      summary: Eliminar regla
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Regla eliminada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
...
  /ollama/config:
    get:
      operationId: getOllamaConfig
      tags: [ollama]
      summary: Obtener configuración de Ollama
      responses:
        "200":
          description: Configuración de Ollama
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OllamaConfig"
    put:
      operationId: updateOllamaConfig
      tags: [ollama]
      summary: Actualizar configuración de Ollama
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OllamaConfigUpdate"
      responses:
        "200":
          description: Configuración actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OllamaConfig"
  /ollama/test:
    post:
      operationId: testOllamaConnection
      tags: [ollama]
      summary: Probar conexión con Ollama
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OllamaTestRequest"
      responses:
        "200":
          description: Resultado del test
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OllamaTestResult"
...
components:
  schemas:
...
    Contact:
      type: object
      properties:
        id:
          type: string
        phone:
          type: string
        name:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        notes:
          type: string
          nullable: true
        stage:
          type: string
          enum: [lead, prospect, customer, vip]
        totalPurchases:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, phone, stage, totalPurchases, createdAt, updatedAt]
    ContactList:
      type: object
      properties:
        contacts:
          type: array
          items:
            $ref: "#/components/schemas/Contact"
        total:
          type: number
      required: [contacts, total]
    CreateContactRequest:
      type: object
      properties:
        phone:
          type: string
        name:
          type: string
        email:
          type: string
        tags:
          type: array
          items:
            type: string
        notes:
          type: string
        stage:
          type: string
          enum: [lead, prospect, customer, vip]
      required: [phone]
    UpdateContactRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        tags:
          type: array
          items:
            type: string
        notes:
          type: string
        stage:
          type: string
          enum: [lead, prospect, customer, vip]
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        category:
          type: string
          nullable: true
        inStock:
          type: boolean
        imageUrl:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, price, inStock, createdAt, updatedAt]
    ProductList:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: number
      required: [products, total]
    CreateProductRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        inStock:
          type: boolean
        imageUrl:
          type: string
      required: [name, price]
    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        inStock:
          type: boolean
        imageUrl:
          type: string
    AutomationRule:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        trigger:
          type: string
          enum: [keyword, first_message, inactivity, purchase]
        triggerValue:
          type: string
          nullable: true
        action:
          type: string
          enum: [send_message, send_catalog, tag_contact, assign_stage, notify_human]
        actionValue:
          type: string
          nullable: true
        enabled:
          type: boolean
        priority:
          type: number
        createdAt:
          type: string
      required: [id, name, trigger, action, enabled, priority, createdAt]
    AutomationRuleList:
      type: object
      properties:
        rules:
          type: array
          items:
            $ref: "#/components/schemas/AutomationRule"
        total:
          type: number
      required: [rules, total]
    CreateAutomationRuleRequest:
      type: object
      properties:
        name:
          type: string
        trigger:
          type: string
          enum: [keyword, first_message, inactivity, purchase]
        triggerValue:
          type: string
        action:
          type: string
          enum: [send_message, send_catalog, tag_contact, assign_stage, notify_human]
        actionValue:
          type: string
        enabled:
          type: boolean
        priority:
          type: number
      required: [name, trigger, action]
    UpdateAutomationRuleRequest:
      type: object
      properties:
        name:
          type: string
        trigger:
          type: string
          enum: [keyword, first_message, inactivity, purchase]
        triggerValue:
          type: string
        action:
          type: string
          enum: [send_message, send_catalog, tag_contact, assign_stage, notify_human]
        actionValue:
          type: string
        enabled:
          type: boolean
        priority:
          type: number
...
    OllamaConfig:
      type: object
      properties:
        url:
          type: string
        model:
          type: string
        temperature:
          type: number
        maxTokens:
          type: number
        connected:
          type: boolean
      required: [url, model, temperature, maxTokens, connected]
    OllamaConfigUpdate:
      type: object
      properties:
        url:
          type: string
        model:
          type: string
        temperature:
          type: number
        maxTokens:
          type: number
    OllamaTestRequest:
      type: object
      properties:
        message:
          type: string
        model:
          type: string
      required: [message]
    OllamaTestResult:
      type: object
      properties:
        response:
          type: string
        model:
          type: string
        latencyMs:
          type: number
        success:
          type: boolean
      required: [response, model, latencyMs, success]
...
    DeleteResult:
      type: object
      properties:
        success:
          type: boolean
        id:
          type: string
      required: [success, id]