# lib/api-zod/src/generated/types/createProductRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API para Bot de Ventas WhatsApp con OpenClaw y Ollama
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductRequest {
  name: string;
  description?: string;
  price: number;
  category?: string;
  inStock?: boolean;
  imageUrl?: string;
}