# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API para Bot de Ventas WhatsApp con OpenClaw y Ollama
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: string;
  name: string;
  description?: string | null;
  price: number;
  category?: string | null;
  inStock: boolean;
  imageUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}